Refactor reevaluate logic
