+++ hooks/useCurrentSpaceId.tsx
@@ -1,9 +1,6 @@
 import type { Dispatch, ReactNode, SetStateAction } from 'react';
-import { useEffect, createContext, useContext, useMemo, useState } from 'react';
+import { createContext, useContext, useMemo, useState } from 'react';
 
-import charmClient from 'charmClient';
-import { useUser } from 'hooks/useUser';
-import { useWeb3AuthSig } from 'hooks/useWeb3AuthSig';
 // keep track of the focused page (may be different from what's in the URL or header)
 
 type ICurrentSpaceContext = {
@@ -18,28 +15,6 @@ export const CurrentSpaceContext = createContext<Readonly<ICurrentSpaceContext>>
 
 export function CurrentSpaceProvider({ children }: { children: ReactNode }) {
   const [currentSpaceId, setCurrentSpaceId] = useState<string>('');
-  const { user, refreshUser } = useUser();
-  const { getStoredSignature, account } = useWeb3AuthSig();
-  useEffect(() => {
-    const reevaluateRoles = async () => {
-      if (user?.id && currentSpaceId && account) {
-        const authSig = getStoredSignature();
-        if (!authSig) return;
-
-        const newRoles = await charmClient.tokenGates.reevaluateRoles({
-          spaceId: currentSpaceId,
-          userId: user.id,
-          authSig
-        });
-
-        if (newRoles.length) {
-          refreshUser();
-        }
-      }
-    };
-
-    reevaluateRoles();
-  }, [account, currentSpaceId, user?.id]);
 
   const value: ICurrentSpaceContext = useMemo(
     () => ({
